Entity classes are back
