+++ src/DigitalLibrary.sln
@@ -120,8 +120,6 @@ ProjectSection(SolutionItems) = preProject
 	Rulesets\xunit.ruleset = Rulesets\xunit.ruleset
 EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MasterData.BusinessLogic.Implementations.Tests.Entities", "MasterData\MasterData.BusinessLogic.Implementations.Tests.Entities\MasterData.BusinessLogic.Implementations.Tests.Entities.csproj", "{AF0AE2A5-B2AD-451C-9FBB-57E745F87E52}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MasterDataTestHelper", "Utils\MasterDataTestHelper\MasterDataTestHelper.csproj", "{9802236C-BED2-4F3E-961C-3541B084CD8D}"
 EndProject
 Global
@@ -306,10 +304,6 @@ Global
 		{435334E1-47EC-4B68-B84D-51EA5C7D7225}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{435334E1-47EC-4B68-B84D-51EA5C7D7225}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{435334E1-47EC-4B68-B84D-51EA5C7D7225}.Release|Any CPU.Build.0 = Release|Any CPU
-		{AF0AE2A5-B2AD-451C-9FBB-57E745F87E52}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AF0AE2A5-B2AD-451C-9FBB-57E745F87E52}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AF0AE2A5-B2AD-451C-9FBB-57E745F87E52}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AF0AE2A5-B2AD-451C-9FBB-57E745F87E52}.Release|Any CPU.Build.0 = Release|Any CPU
 		{9802236C-BED2-4F3E-961C-3541B084CD8D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{9802236C-BED2-4F3E-961C-3541B084CD8D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9802236C-BED2-4F3E-961C-3541B084CD8D}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -363,7 +357,6 @@ Global
 		{A5DBA3B7-EACB-48DB-BEC7-C3FC551CCBA6} = {51AD6AE7-80CE-4882-8F97-D22DD30628B8}
 		{26BABB9A-5C6F-4B0D-A2B0-AAC33149ECF7} = {51AD6AE7-80CE-4882-8F97-D22DD30628B8}
 		{435334E1-47EC-4B68-B84D-51EA5C7D7225} = {51AD6AE7-80CE-4882-8F97-D22DD30628B8}
-		{AF0AE2A5-B2AD-451C-9FBB-57E745F87E52} = {2E768E9C-7D0C-47F3-A42A-7B7E0983A010}
 		{9802236C-BED2-4F3E-961C-3541B084CD8D} = {5DC05A13-9489-49BE-8335-C5A6E766CDF8}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
