+++ src/MasterData/MasterData.BusinessLogic.Implementations.Tests/MasterData.BusinessLogic.Implementations.Tests.csproj
@@ -46,7 +46,6 @@
         <ProjectReference Include="..\..\Utils\Guards\Guards.csproj"/>
         <ProjectReference Include="..\..\Utils\IntegrationTestFactories\IntegrationTestFactories.csproj"/>
         <ProjectReference Include="..\..\Utils\MasterDataTestHelper\MasterDataTestHelper.csproj"/>
-        <ProjectReference Include="..\MasterData.BusinessLogic.Implementations.Tests.Entities\MasterData.BusinessLogic.Implementations.Tests.Entities.csproj"/>
         <ProjectReference Include="..\MasterData.BusinessLogic.Implementations\MasterData.BusinessLogic.Implementations.csproj"/>
         <ProjectReference Include="..\MasterData.BusinessLogic.Interfaces\MasterData.BusinessLogic.Interfaces.csproj"/>
         <ProjectReference Include="..\MasterData.Validators.TestData\MasterData.Validators.TestData.csproj"/>
