+++ src/MasterData/MasterData.BusinessLogic.Implementations.Tests/StepDefinitions/DifferenceBetweenActiveAmountBeforeLogicalDeleteAndAfterIs.cs
@@ -1,0 +1,37 @@
+namespace DigitalLibrary.MasterData.BusinessLogic.Implementations.Tests.StepDefinitions
+{
+    using System.Diagnostics.CodeAnalysis;
+
+    using DigitalLibrary.Utils.Guards;
+
+    using FluentAssertions;
+
+    using TechTalk.SpecFlow;
+    using TechTalk.SpecFlow.Assist;
+
+    public partial class StepDefinitions
+    {
+        [Then(@"difference between lists is")]
+        public void DifferenceBetweenActiveAmountBeforeLogicalDeleteAndAfterIs(Table table)
+        {
+            Check.IsNotNull(table);
+            DifferenceBetweenActiveAmountBeforeLogicalDeleteAndAfterIsEntity instance =
+                table.CreateInstance<DifferenceBetweenActiveAmountBeforeLogicalDeleteAndAfterIsEntity>();
+
+            int amountBefore = (int) _scenarioContext[instance.FirstResultKey];
+            int amountAfter = (int) _scenarioContext[instance.SecondResultKey];
+
+            int result = amountBefore - amountAfter;
+            result.Should().Be(instance.ExpectedDiff);
+        }
+    }
+
+    internal class DifferenceBetweenActiveAmountBeforeLogicalDeleteAndAfterIsEntity
+    {
+        public int ExpectedDiff { get; set; }
+
+        public string FirstResultKey { get; set; }
+
+        public string SecondResultKey { get; set; }
+    }
+}
