+++ src/MasterData/MasterData.BusinessLogic.Implementations.Tests/StepDefinitions/DimensionStructureIsLogicallyDeleted.cs
@@ -1,0 +1,48 @@
+namespace DigitalLibrary.MasterData.BusinessLogic.Implementations.Tests.StepDefinitions
+{
+    using System;
+    using System.Threading.Tasks;
+
+    using DigitalLibrary.MasterData.DomainModel;
+    using DigitalLibrary.Utils.Guards;
+
+    using TechTalk.SpecFlow;
+    using TechTalk.SpecFlow.Assist;
+
+    public partial class StepDefinitions
+    {
+        [When(@"DimensionStructure is logically deleted")]
+        public async Task DimensionStructureIsLogicallyDeleted(Table table)
+        {
+            Check.IsNotNull(table);
+            DimensionStructureIsLogicallyDeletedEntity instance = table
+               .CreateInstance<DimensionStructureIsLogicallyDeletedEntity>();
+
+            DimensionStructure toBeDeleted = _scenarioContext[instance.Key] as DimensionStructure;
+            Check.IsNotNull(toBeDeleted);
+
+            try
+            {
+                DimensionStructure dimensionStructure = new DimensionStructure
+                {
+                    Id = toBeDeleted.Id,
+                };
+                await _masterDataBusinessLogic.MasterDataDimensionStructureBusinessLogic
+                   .DeleteLogicallyAsync(dimensionStructure)
+                   .ConfigureAwait(false);
+                _scenarioContext.Add(instance.ResultKey, SUCCESS);
+            }
+            catch (Exception e)
+            {
+                _scenarioContext.Add(instance.ResultKey, e);
+            }
+        }
+    }
+
+    internal class DimensionStructureIsLogicallyDeletedEntity
+    {
+        public string ResultKey { get; set; }
+
+        public string Key { get; set; }
+    }
+}
