+++ src/MasterData/MasterData.BusinessLogic.Implementations.Tests/StepDefinitions/DimensionStructureIsRequested.cs
@@ -1,0 +1,44 @@
+namespace DigitalLibrary.MasterData.BusinessLogic.Implementations.Tests.StepDefinitions
+{
+    using System;
+    using System.Threading.Tasks;
+
+    using DigitalLibrary.MasterData.DomainModel;
+    using DigitalLibrary.Utils.Guards;
+
+    using TechTalk.SpecFlow;
+    using TechTalk.SpecFlow.Assist;
+
+    public partial class StepDefinitions
+    {
+        [When(@"DimensionStructure is requested")]
+        public async Task DimensionStructureIsRequested(Table table)
+        {
+            Check.IsNotNull(table);
+
+            DimensionStructureIsRequestedEntity instance = table.CreateInstance<DimensionStructureIsRequestedEntity>();
+
+            DimensionStructure dimensionStructure = _scenarioContext[instance.Key] as DimensionStructure;
+
+            try
+            {
+                DimensionStructure dimensionStructureResult = await _masterDataBusinessLogic
+                   .MasterDataDimensionStructureBusinessLogic
+                   .GetDimensionStructureByIdAsync(dimensionStructure)
+                   .ConfigureAwait(false);
+                _scenarioContext.Add(instance.ResultKey, dimensionStructureResult);
+            }
+            catch (Exception e)
+            {
+                _scenarioContext.Add(instance.ResultKey, e);
+            }
+        }
+    }
+
+    internal class DimensionStructureIsRequestedEntity
+    {
+        public string Key { get; set; }
+
+        public string ResultKey { get; set; }
+    }
+}
