+++ src/MasterData/MasterData.BusinessLogic.Implementations.Tests/StepDefinitions/DimensionStructurePropertyIsNotNull.cs
@@ -1,0 +1,48 @@
+namespace DigitalLibrary.MasterData.BusinessLogic.Implementations.Tests.StepDefinitions
+{
+    using DigitalLibrary.MasterData.DomainModel;
+    using DigitalLibrary.Utils.Guards;
+    using DigitalLibrary.Utils.MasterDataTestHelper;
+
+    using FluentAssertions;
+
+    using TechTalk.SpecFlow;
+    using TechTalk.SpecFlow.Assist;
+
+    public partial class StepDefinitions
+    {
+        [Then(@"DimensionStructure property is not null")]
+        public void DimensionStructurePropertyIsNotNull(Table table)
+        {
+            Check.IsNotNull(table);
+
+            DimensionStructurePropertyIsNotNullEntity instance = table
+               .CreateInstance<DimensionStructurePropertyIsNotNullEntity>();
+
+            DimensionStructure result = _scenarioContext[instance.Key] as DimensionStructure;
+            Check.IsNotNull(result);
+
+            switch (instance.PropertyName)
+            {
+                case DimensionStructurePropertiesStruct.Name:
+                    result.Name.Should().NotBeNull();
+                    break;
+
+                case DimensionStructurePropertiesStruct.Desc:
+                    result.Desc.Should().NotBeNull();
+                    break;
+
+                default:
+                    string msg = $"{instance.PropertyName} doesn't exist or not case covered.";
+                    throw new MasterDataStepDefinitionException(msg);
+            }
+        }
+    }
+
+    internal class DimensionStructurePropertyIsNotNullEntity
+    {
+        public string Key { get; set; }
+
+        public string PropertyName { get; set; }
+    }
+}
