+++ src/MasterData/MasterData.BusinessLogic.Implementations.Tests/StepDefinitions/SourceFormatIsRequestedWithRootDimensionStructure.cs
@@ -1,0 +1,36 @@
+namespace DigitalLibrary.MasterData.BusinessLogic.Implementations.Tests.StepDefinitions
+{
+    using System.Threading.Tasks;
+
+    using DigitalLibrary.MasterData.DomainModel;
+    using DigitalLibrary.Utils.Guards;
+
+    using TechTalk.SpecFlow;
+    using TechTalk.SpecFlow.Assist;
+
+    public partial class StepDefinitions
+    {
+        [When(@"SourceFormat is requested with root DimensionStructure")]
+        public async Task SourceFormatIsRequestedWithRootDimensionStructure(Table table)
+        {
+            SourceFormatIsRequestedWithRootDimensionStructureEntity instance = table
+               .CreateInstance<SourceFormatIsRequestedWithRootDimensionStructureEntity>();
+
+            DomainModel.SourceFormat withoutTree = _scenarioContext[instance.Key] as SourceFormat;
+
+            DomainModel.SourceFormat requestedWithTree = await _masterDataBusinessLogic
+               .MasterDataSourceFormatBusinessLogic
+               .GetSourceFormatByIdWithRootDimensionStructureAsync(withoutTree)
+               .ConfigureAwait(false);
+
+            _scenarioContext.Add(instance.ResultKey, requestedWithTree);
+        }
+    }
+
+    internal class SourceFormatIsRequestedWithRootDimensionStructureEntity
+    {
+        public string Key { get; set; }
+
+        public string ResultKey { get; set; }
+    }
+}
