+++ src/MasterData/MasterData.BusinessLogic.Implementations.Tests/StepDefinitions/SourceFormatRelatedOperationThrowsException.cs
@@ -1,22 +1,27 @@
-namespace DigitalLibrary.MasterData.BusinessLogic.Implementations.Tests.stepdefs
+namespace DigitalLibrary.MasterData.BusinessLogic.Implementations.Tests.StepDefinitions
 {
     using DigitalLibrary.MasterData.BusinessLogic.Implementations.SourceFormat;
-    using DigitalLibrary.MasterData.BusinessLogic.Implementations.Tests.Entities;
 
     using FluentAssertions;
 
     using TechTalk.SpecFlow;
     using TechTalk.SpecFlow.Assist;
 
-    public partial class StepDefs
+    public partial class StepDefinitions
     {
         [Then(@"SourceFormat related operation throws exception")]
         public void SourceFormatRelatedOperationThrowsException(Table table)
         {
-            ExceptionWasThrownEntity instance = table.CreateInstance<ExceptionWasThrownEntity>();
+            SourceFormatRelatedOperationThrowsExceptionEntity instance = table
+               .CreateInstance<SourceFormatRelatedOperationThrowsExceptionEntity>();
 
             var exception = _scenarioContext[instance.ResultKey];
             exception.Should().BeOfType<MasterDataBusinessLogicSourceFormatDatabaseOperationException>();
         }
+    }
+
+    internal class SourceFormatRelatedOperationThrowsExceptionEntity
+    {
+        public string ResultKey { get; set; }
     }
 }
