+++ src/MasterData/MasterData.BusinessLogic.Implementations.Tests/StepDefinitions/SourceFormatResultsRootDimensionStructureIdIsNotZero.cs
@@ -1,0 +1,23 @@
+namespace DigitalLibrary.MasterData.BusinessLogic.Implementations.Tests.StepDefinitions
+{
+    using DigitalLibrary.MasterData.DomainModel;
+    using DigitalLibrary.Utils.Guards;
+
+    using FluentAssertions;
+
+    using TechTalk.SpecFlow;
+
+    public partial class StepDefinitions
+    {
+        [Then(@"'(.*)' SourceFormat result's RootDimensionStructure Id property is not zero")]
+        public void SourceFormatResultsRootDimensionStructureIdIsNotZero(string resultKey)
+        {
+            SourceFormat result = _scenarioContext[resultKey] as SourceFormat;
+            Check.IsNotNull(result);
+
+            result.SourceFormatDimensionStructureNode.DimensionStructureNode.Id
+               .Should()
+               .NotBe(0);
+        }
+    }
+}
