+++ src/MasterData/MasterData.BusinessLogic.Implementations.Tests/StepDefinitions/SourceFormatResultsRootDimensionStructurePropertyEqualsTo.cs
@@ -1,0 +1,57 @@
+namespace DigitalLibrary.MasterData.BusinessLogic.Implementations.Tests.StepDefinitions
+{
+    using DigitalLibrary.MasterData.DomainModel;
+    using DigitalLibrary.Utils.MasterDataTestHelper;
+
+    using FluentAssertions;
+
+    using TechTalk.SpecFlow;
+    using TechTalk.SpecFlow.Assist;
+
+    public partial class StepDefinitions
+    {
+        [Then(@"SourceFormat result's RootDimensionStructure property equals to")]
+        public void SourceFormatResultsRootDimensionStructurePropertyEqualsTo(Table table)
+        {
+            SourceFormatResultsRootDimensionStructurePropertyEqualsToEntity instance =
+                table.CreateInstance<SourceFormatResultsRootDimensionStructurePropertyEqualsToEntity>();
+
+            DomainModel.SourceFormat result = _scenarioContext[instance.Key] as SourceFormat;
+            DomainModel.DimensionStructure comparedTo = _scenarioContext[instance.EqualsTo] as DimensionStructure;
+
+            switch (instance.PropertyName)
+            {
+                case DimensionStructurePropertiesStruct.Name:
+                    result.SourceFormatDimensionStructureNode.DimensionStructureNode.DimensionStructure.Name
+                       .Should()
+                       .Be(comparedTo.Name);
+                    break;
+
+                case DimensionStructurePropertiesStruct.Desc:
+                    result.SourceFormatDimensionStructureNode.DimensionStructureNode.DimensionStructure.Desc
+                       .Should()
+                       .Be(comparedTo.Desc);
+                    break;
+
+                case DimensionStructurePropertiesStruct.IsActive:
+                    result.SourceFormatDimensionStructureNode.DimensionStructureNode.DimensionStructure.IsActive
+                       .Should()
+                       .Be(comparedTo.IsActive);
+                    break;
+
+                default:
+                    throw new MasterDataStepDefinitionException(
+                        nameof(SourceFormatResultsRootDimensionStructurePropertyEqualsTo));
+            }
+        }
+    }
+
+    internal class SourceFormatResultsRootDimensionStructurePropertyEqualsToEntity
+    {
+        public string EqualsTo { get; set; }
+
+        public string Key { get; set; }
+
+        public string PropertyName { get; set; }
+    }
+}
