+++ src/MasterData/MasterData.BusinessLogic.Implementations.Tests/StepDefinitions/ThenDimensionStructureTreeContainsGivenDimensionStructure.cs
@@ -1,0 +1,44 @@
+namespace DigitalLibrary.MasterData.BusinessLogic.Implementations.Tests.StepDefinitions
+{
+    using System.Linq;
+    using System.Threading.Tasks;
+
+    using DigitalLibrary.MasterData.DomainModel;
+    using DigitalLibrary.Utils.Guards;
+
+    using FluentAssertions;
+
+    using TechTalk.SpecFlow;
+    using TechTalk.SpecFlow.Assist;
+
+    public partial class StepDefinitions
+    {
+        [Then(@"DimensionStructure tree contains given DimensionStructure")]
+        public async Task ThenDimensionStructureTreeContainsGivenDimensionStructure(Table table)
+        {
+            Check.IsNotNull(table);
+
+            ThenDimensionStructureTreeContainsGivenDimensionStructureEntity instance = table
+               .CreateInstance<ThenDimensionStructureTreeContainsGivenDimensionStructureEntity>();
+
+            SourceFormat sourceFormat = _scenarioContext[instance.SourceFormatKey] as SourceFormat;
+            DimensionStructure dimensionStructure = _scenarioContext[instance.ContainedDimensionStructureKey]
+                as DimensionStructure;
+
+            SourceFormat withAllInvolvedNodes = await _masterDataBusinessLogic.MasterDataSourceFormatBusinessLogic
+               .GetSourceFormatByIdWithAllDimensionStructuresAndNodesAsync(sourceFormat)
+               .ConfigureAwait(false);
+
+            withAllInvolvedNodes.DimensionStructureNodes
+               .FirstOrDefault(c => c.DimensionStructure.Id == dimensionStructure.Id)
+               .Should().NotBeNull();
+        }
+    }
+
+    internal class ThenDimensionStructureTreeContainsGivenDimensionStructureEntity
+    {
+        public string SourceFormatKey { get; set; }
+
+        public string ContainedDimensionStructureKey { get; set; }
+    }
+}
