+++ src/MasterData/MasterData.BusinessLogic.Implementations.Tests/StepDefinitions/ThereIsASavedDimensionStructureDomainObject.cs
@@ -1,0 +1,50 @@
+namespace DigitalLibrary.MasterData.BusinessLogic.Implementations.Tests.StepDefinitions
+{
+    using System;
+    using System.Threading.Tasks;
+
+    using DigitalLibrary.MasterData.DomainModel;
+    using DigitalLibrary.Utils.MasterDataTestHelper.Tools;
+
+    using TechTalk.SpecFlow;
+    using TechTalk.SpecFlow.Assist;
+
+    public partial class StepDefinitions
+    {
+        [Given(@"there is a saved DimensionStructure domain object")]
+        public async Task ThereIsASavedDimensionStructureDomainObject(Table table)
+        {
+            ThereIsASavedDimensionStructureDomainObjectEntity instance = table
+               .CreateInstance<ThereIsASavedDimensionStructureDomainObjectEntity>();
+            DomainModel.DimensionStructure dimensionStructure = _masterDataTestHelper
+               .DimensionStructureFactory
+               .Create(instance);
+
+            try
+            {
+                DimensionStructure result = await _masterDataBusinessLogic
+                   .MasterDataDimensionStructureBusinessLogic
+                   .AddAsync(dimensionStructure)
+                   .ConfigureAwait(false);
+                _scenarioContext.Add(instance.ResultKey, result);
+            }
+            catch (Exception e)
+            {
+                _scenarioContext.Add(instance.ResultKey, e);
+            }
+        }
+    }
+
+    internal class ThereIsASavedDimensionStructureDomainObjectEntity : IDimensionStructureDomainObject
+    {
+        public string Desc { get; set; }
+
+        public int IsActive { get; set; }
+
+        public string Key { get; set; }
+
+        public string ResultKey { get; set; }
+
+        public string Name { get; set; }
+    }
+}
