+++ src/Rulesets/StyleCopAnalyzers.ruleset
@@ -165,18 +165,18 @@
     </Rules>
 
     <Rules AnalyzerId="StyleCop.Analyzers" RuleNamespace="StyleCop.Analyzers.DocumentationRules">
-        <Rule Id="SA1600"  Action="Warning" />          <!-- Elements should be documented -->
-        <Rule Id="SA1601"  Action="Warning" />          <!-- Partial elements should be documented -->
-        <Rule Id="SA1602"  Action="Warning" />          <!-- Enumeration items should be documented -->
-        <Rule Id="SA1603"  Action="Warning" />             <!-- Documentation should contain valid XML -->
-        <Rule Id="SA1604"  Action="Warning" />          <!-- Element documentation should have summary -->
-        <Rule Id="SA1605"  Action="Warning" />          <!-- Partial element documentation should have summary -->
-        <Rule Id="SA1606"  Action="Warning" />          <!-- Element documentation should have summary text -->
-        <Rule Id="SA1607"  Action="Warning" />          <!-- Partial element documentation should have summary text -->
-        <Rule Id="SA1608"  Action="Warning" />          <!-- Element documentation should not have default summary -->
-        <Rule Id="SA1609"  Action="Warning" />             <!-- Property documentation should have value -->
-        <Rule Id="SA1610"  Action="Warning" />          <!-- Property documentation should have value text -->
-        <Rule Id="SA1611"  Action="Warning" />          <!-- Element parameters should be documented -->
+        <Rule Id="SA1600" Action="Warning"/>          <!-- Elements should be documented -->
+        <Rule Id="SA1601" Action="None"/>          <!-- Partial elements should be documented -->
+        <Rule Id="SA1602" Action="Warning"/>          <!-- Enumeration items should be documented -->
+        <Rule Id="SA1603" Action="Warning"/>             <!-- Documentation should contain valid XML -->
+        <Rule Id="SA1604" Action="Warning"/>          <!-- Element documentation should have summary -->
+        <Rule Id="SA1605" Action="Warning"/>          <!-- Partial element documentation should have summary -->
+        <Rule Id="SA1606" Action="Warning"/>          <!-- Element documentation should have summary text -->
+        <Rule Id="SA1607" Action="Warning"/>          <!-- Partial element documentation should have summary text -->
+        <Rule Id="SA1608" Action="Warning"/>          <!-- Element documentation should not have default summary -->
+        <Rule Id="SA1609" Action="Warning"/>             <!-- Property documentation should have value -->
+        <Rule Id="SA1610" Action="Warning"/>          <!-- Property documentation should have value text -->
+        <Rule Id="SA1611" Action="Warning"/>          <!-- Element parameters should be documented -->
         <Rule Id="SA1612"  Action="Warning" />          <!-- Element parameter documentation should match element parameters -->
         <Rule Id="SA1613"  Action="Warning" />          <!-- Element parameter documentation should declare parameter name -->
         <Rule Id="SA1614"  Action="Warning" />          <!-- Element parameter documentation should have text -->
