+++ src/Utils/MasterDataTestHelper/MasterDataTestHelper.csproj
@@ -8,7 +8,6 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\..\MasterData\MasterData.BusinessLogic.Implementations.Tests.Entities\MasterData.BusinessLogic.Implementations.Tests.Entities.csproj"/>
         <ProjectReference Include="..\..\MasterData\MasterData.DomainModel\MasterData.DomainModel.csproj"/>
         <ProjectReference Include="..\Guards\Guards.csproj"/>
     </ItemGroup>
