+++ src/Utils/MasterDataTestHelper/Tools/SourceFormatFactory.cs
@@ -3,7 +3,6 @@ namespace DigitalLibrary.Utils.MasterDataTestHelper.Tools
     using System;
     using System.Threading.Tasks;
 
-    using DigitalLibrary.MasterData.BusinessLogic.Implementations.Tests.Entities;
     using DigitalLibrary.MasterData.DomainModel;
     using DigitalLibrary.Utils.Guards;
 
@@ -22,7 +21,8 @@ namespace DigitalLibrary.Utils.MasterDataTestHelper.Tools
         }
 
         /// <inheritdoc/>
-        public SourceFormat Create(ThereIsASourceFormatDomainObjectEntity instance)
+        public SourceFormat Create<T>(T instance)
+            where T : ISourceFormatDomainObject
         {
             Check.IsNotNull(instance);
